+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+--------------------
+##### Added
+- Count-non-empty() support.
+
 [1.5.0] - 2017-04-25 
 --------------------
 ##### Added